Extract image layer graph snapshot writer
